added Window Handler and added release resource from rerources manager

+++ include/Math3D.h
@@ -831,7 +831,7 @@ namespace Easy2D{
 		static DFORCEINLINE float torad(float deg) { return deg*PIOVER180; }
 		static DFORCEINLINE float todeg(float rad) { return rad*G180OVERPI; }
         static DFORCEINLINE float normaliseOrientation(float rot){
-                rot=std::fmodf(rot,(float)Math::PI2);
+                rot=std::fmod(rot,(float)Math::PI2);
                 return rot<0 ? rot+=Math::PI2 : rot;
         }
 		//fast swap
